// src/pages/ielts-prediction/ui/types.ts
export interface PracticeLibraryBannerConfig {
  listening: {
    title: string;
    description: {
      line1: string;
      line2: string;
      line3: string;
    };
    backgroundColor: string;
    button: {
      text: string;
      link: string;
    };
  };
  reading: {
    title: string;
    description: {
      line1: string;
      line2: string;
      line3: string;
    };
    backgroundColor: string;
    button: {
      text: string;
      link: string;
    };
  };
}